- added pytest-timeout to requirements.txt - added conftest.py for global functions/ fixtures - no more deprecated package imp
