+++ requirements.txt
@@ -15,4 +15,4 @@ peigen>=0.0.9
 autograd>=1.3
 lmfit>=1.0.1
 pre-commit
-
+pytest-timeout
